+++ src/libmerc/global_config.h
@@ -190,6 +190,7 @@ public:
             { "ssdp",                   false },
             { "ssh",                    false },
             { "stun",                   false },
+            { "syslog",                 false },
             { "tacacs",                 false },
             { "tcp",                    false },
             { "tcp.message",            false },