+++ src/libmerc/pkt_proc_util.h
@@ -85,6 +85,7 @@ class gre_header;
 class geneve;
 class ip_encapsulation;
 class vxlan;
+class syslog;
 
 using protocol = std::variant<std::monostate,
                               http_request,                      // start of tcp protocols
@@ -136,7 +137,8 @@ using protocol = std::variant<std::monostate,
                               esp,
                               ike::packet,
                               rfb::protocol_version_handshake,
-                              tacacs::packet
+                              tacacs::packet,
+                              syslog
                               >;
 using encapsulation = std::variant<std::monostate,
                               gre_header,